Create MergeSort
This is a simple merging algorithm where I merge two integer arrays together. First I wrote a method to sort an array in ascending order, then another method to merge the two arrays  together, as well as a method to print the array, and lastly a method to scramble an array.