Create Banking Application v1
This is a simple banking application that allows users to sign into an account that already exists or create a new account where they'll be asked to enter their first and last name, as well as a password.  After either signing into an existing account or creating an account users are then presented with a list of options to select from.

These options are to create another account, update certain information of the account that is logged in, withdraw or deposit funds into an account, transfer between saving and checking account, send funds to another person, list all accounts associated with the bank, logging out, and exiting the program.

The way this program works is once it's ran it'll check if there's a text file named "AccountList.txt" if there isn't then it'll create one, and if there is then it'll read from that text file and populate an ArrayList. After every action the ArrayList is updated to keep track of changes made throughout the run of the program. Once the user chooses the "Exit" option the ArrayList which holds all the user options is then written into the text file and overwrites whatever was in that text file to avoid duplications.