Create Encryption
I wrote a program that prompts the user for a message they wish to encrypt. Once it receives the message it parses through calling the encryption method and storing it into an array list.

To encrypt it we store the string into a character array and randomly swap the characters while keeping track of what positions have been swapped. While storing the positions that have been swapped into an array list it also adds randomly generated numbers that act as red herrings. Then we store the ascii values of the characters in order, while also adding red herring values that are randomly generated withing the average character's ascii values of the original characters. Lastly, it stores the character size of the string multiplied by 4.

I also included the decryption formula for the message to recreate it.