Add files via upload
This program will generate a password of varying complexity on a scale of 1 to 4, 1 being very simple to 4 being more complex. The user will be prompted to give a purpose to the password, which will be stored into a string variable. Then the user will be asked to chose the length of the password they want with 8 being the minimum and then the complexity of the password, 1 being only lower case letters, 2 being upper and lower case letters, 3 being upper and lower case letters as well as numbers, and 4 being upper and lower case letters, numbers, and special symbols.

Then once the password is generated it will be stored into a text-file named "passwordGeneratorStorage.txt" and it will store the purpose of the password, the password itself, and the date and time it was created.

In future updates I wish to create a login system that will keep track of different users who use to program and restrict access to private information unless the user logs into their respective account.